refactor: name in platform and avoid cycle dep with helpers
