+++ libs/ledger-live-common/src/helpers.ts
@@ -9,3 +9,9 @@ export function includes<T extends U, U>(array: ReadonlyArray<T>, element: U): e
 export function objectKeysType<Type extends object>(value: Type): Array<keyof Type> {
   return Object.keys(value) as Array<keyof Type>;
 }
+
+export function reverseRecord<T extends PropertyKey, U extends PropertyKey>(
+  input: Record<T, U>,
+): Record<U, T> {
+  return Object.fromEntries(Object.entries(input).map(([key, value]) => [value, key]));
+}
